+++ src/mollie/index.ts
@@ -2,7 +2,12 @@
 export { MollieService } from './services/mollie_service'
 export { CustomerService } from './services/customer_service'
 export { InvoiceService } from './services/invoice_service'
+export { InvoiceAuditService } from './services/invoice_audit_service'
 export { SubscriptionService } from './services/subscription_service'
+
+// ORMs (audit log)
+export { InvoiceStatusLogOrm } from './orm/invoice_status_log_orm'
+export { PaymentStatusLogOrm } from './orm/payment_status_log_orm'
 
 // Public types (entities, inputs, outputs)
 export * from './types/mollie_types'